Stupidity

+++ AoC_2021/IChallenge.cs
@@ -8,5 +8,5 @@ namespace AoC2021;
 
 public interface IChallenge
 {
-    public int Answer { get; }
+    public long Answer { get; }
 }
